Trait cry::primitive::point::DisLogPoint

pub trait DisLogPoint: Clone + Bytes + Debug {
    type Scalar: ScalarNumber;

    const SIZE: usize;

    fn zero() -> Self;

    fn one() -> Self;

    fn basepoint() -> Self;

    fn add(&self, rhs: &Self) -> Self;

    fn mul(&self, rhs: &Self::Scalar) -> Self;

    fn neg(&self) -> Self;

    fn eq(&self, o: &Self) -> bool;

    fn sub(&self, rhs: &Self) -> Self
    where
        Self: Sized,
    { ... }
}

Point operator for this curve on finite finite $G_p$

Associated Types

type Scalar: ScalarNumber

This point type’s associated scalar type.

Beacuse point can’t multiply by point, We need use scalar multiply for point.

Loading content...
Associated Constants

const SIZE: usize

Point size in bytes.

Loading content...
Required methods

fn zero() -> Self

Get a point means zero.

fn one() -> Self

Get a point means one.

fn basepoint() -> Self

Get a base point for this curve.

fn add(&self, rhs: &Self) -> Self

Add two point.

fn mul(&self, rhs: &Self::Scalar) -> Self

Scalar multiply for point.

fn neg(&self) -> Self

Negative a point.

fn eq(&self, o: &Self) -> bool

Two point is equal.

Loading content...
Provided methods

fn sub(&self, rhs: &Self) -> Self where
Self: Sized,

Subtraction two point.

Loading content...